requirements.txt
orjson>=3.8.0
pillow>=9.0.0
psycopg2-binary>=2.9.5
pyvips>=2.2
requests>=2.28
tqdm>=4.64.0
